backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pymysql==1.1.0
python-dotenv==1.0.0
alembic==1.12.1
pydantic-settings==2.1.0
watchdog==3.0.0
email-validator==2.1.0.post1
fastapi-mail==1.4.1
loguru==0.7.2
python-json-logger==2.0.7

# 开发依赖
pytest==7.4.3
black==23.10.1
isort==5.12.0
flake8==6.1.0
pytest-asyncio==0.21.1 